requirements.txt
discord.py
psycopg2-binary==2.8.6
